requirements.txt
zendriver
beautifulsoup4
requests
psycopg2-binary
